requirements.txt
numpy
matplotlib
scikit-learn
seaborn
pandas
tqdm
transformers
torch
torchvision
fastapi
uvicorn
pillow
streamlit>=1.28.0
python-multipart
